+++ moor_generator/lib/src/model/declarations/declaration.dart
@@ -6,6 +6,7 @@ import 'package:sqlparser/sqlparser.dart';
 part 'columns.dart';
 part 'database.dart';
 part 'tables.dart';
+part 'trigger.dart';
 
 /// Interface for model elements that are declared somewhere.
 abstract class HasDeclaration {