+++ moor_generator/lib/src/model/declarations/trigger.dart
@@ -1,0 +1,16 @@
+part of 'declaration.dart';
+
+abstract class TriggerDeclaration extends Declaration {}
+
+class MoorTriggerDeclaration implements MoorDeclaration, TriggerDeclaration {
+  @override
+  final SourceRange declaration;
+
+  @override
+  final CreateTriggerStatement node;
+
+  MoorTriggerDeclaration._(this.declaration, this.node);
+
+  MoorTriggerDeclaration.fromNodeAndFile(this.node, FoundFile file)
+      : declaration = SourceRange.fromNodeAndFile(node, file);
+}